int -> unsigned long int for sum variables to fix overflow issue

+++ fastqcheck.c
@@ -18,6 +18,8 @@
  * gcc -Wall gripes.
  * Change requested by Richard Durbin, Petr Danecek, to space long
  * numbers rather than rely on fixed width fields.
+ * sum, qsum, psum and pqsum all changed to unsigned long int to fix
+ * overflow error.
  */
 
 #include <stdlib.h>
@@ -35,13 +37,13 @@
 int main (int argc, char **argv)
 {
   int i, j, length, lengthMax = 0, qMax = 0, status ;
-  int nseq = 0;
+  int nseq = 0 ;
   unsigned long int total = 0 ;
   char *seq, *id ;
   unsigned char *qval ;
   FILE *fil ;
-  static int sum[5], qsum[256] ;		/* 0 automatically */
-  static int psum[MAX_LENGTH][5], pqsum[MAX_LENGTH][256], nlen[MAX_LENGTH] ;
+  static unsigned long int sum[5], qsum[256] ;        /* 0 automatically */
+  static unsigned long int psum[MAX_LENGTH][5], pqsum[MAX_LENGTH][256], nlen[MAX_LENGTH] ;
   double erate;
 
   if (argc == 1)
@@ -54,25 +56,27 @@ int main (int argc, char **argv)
   while ( (status=readFastq (fil, dna2indexConv, &seq, &qval, &id, &length)) > 0)
     { ++nseq ; ++nlen[length] ;
       total += length ;
+
       if (length > lengthMax) 
-	{ lengthMax = length ;
-	  if (length > MAX_LENGTH)
-	    { fprintf (stderr, "read %s length = %d longer than MAX_LENGTH = %d; edit and recompile with larger MAX_LENGTH\n", id, length, MAX_LENGTH) ;
-	      exit (EXIT_FAILURE) ;
-	    }
-	}
+        { lengthMax = length ;
+          if (length > MAX_LENGTH)
+            { fprintf (stderr, "read %s length = %d longer than MAX_LENGTH = %d; edit and recompile with larger MAX_LENGTH\n", id, length, MAX_LENGTH) ;
+              exit (EXIT_FAILURE) ;
+            }
+        }
+
       for (i = 0 ; i < length ; ++i)
-	{ ++sum[(int)seq[i]] ; ++psum[i][(int)seq[i]] ;
-	  ++qsum[qval[i]] ; ++pqsum[i][qval[i]] ;
-	  if (qval[i] > qMax) qMax = qval[i] ;
-	}
+        { ++sum[(int)seq[i]] ; ++psum[i][(int)seq[i]] ;
+          ++qsum[qval[i]] ; ++pqsum[i][qval[i]] ;
+          if (qval[i] > qMax) qMax = qval[i] ;
+        }
       free (seq) ; free (qval) ; free (id) ;
     }
 
     if ( status<0 )
-    {
+      {
         exit (EXIT_FAILURE);
-    }
+      }
 
   printf ("%d sequences, %ld total length", nseq, total) ;
   if (nseq)
@@ -81,34 +85,47 @@ int main (int argc, char **argv)
 
   if (total)
     { printf ("Standard deviations at 0.25:  total %5.2f %%, per base %5.2f %%\n", 
-	      100*(sqrt(0.25*(double)total)/total), 100*(sqrt(0.25*(double)nseq)/nseq)) ;
+          100*(sqrt(0.25*(double)total)/total), 100*(sqrt(0.25*(double)nseq)/nseq)) ;
       printf ("            A    C    G    T    N ") ;
       for (i = 0 ; i <= qMax ; ++i) printf (" %3d",i) ;
       printf (" AQ\nTotal  ") ;
       printf ("  %4.1f %4.1f %4.1f %4.1f %4.1f ", 
-	      100*((double)sum[_A]/total), 100*((double)sum[_C]/total),
-	      100*((double)sum[_G]/total), 100*((double)sum[_T]/total),
-	      100*((double)sum[_N]/total)) ;
-      for (erate = j = 0 ; j <= qMax ; ++j) {
-	printf (" %3d",(int)lrint(1000*((double)qsum[j]/total))) ;
-	erate += pow(10, j/-10.0) * qsum[j];
-      }
+          100*((double)sum[_A]/total), 100*((double)sum[_C]/total),
+          100*((double)sum[_G]/total), 100*((double)sum[_T]/total),
+          100*((double)sum[_N]/total)) ;
+
+
+
+
+
+      for (erate = j = 0 ; j <= qMax ; ++j)
+        {
+          printf (" %3d",(int)lrint(1000*((double)qsum[j]/total))) ;
+          erate += pow(10, j/-10.0) * qsum[j];
+        }
+
+
+
+
+
       printf(" %4.1f", -10*log(erate/total)/log(10));
       for (i = 0 ; i < lengthMax ; ++i)
-	{ nseq -= nlen[i] ;
-	  printf ("\nbase %2d", i+1) ;
-	  printf ("  %4.1f %4.1f %4.1f %4.1f %4.1f ",
-		  100*((double)psum[i][_A]/nseq), 100*((double)psum[i][_C]/nseq),
-		  100*((double)psum[i][_G]/nseq), 100*((double)psum[i][_T]/nseq),
-		  100*((double)psum[i][_N]/nseq)) ;
-	  for (erate = j = 0 ; j <= qMax ; ++j) {
-	      printf (" %3d",(int)lrint(1000*((double)pqsum[i][j]/nseq))) ;
-	      erate += pow(10, j/-10.0) * pqsum[i][j];
-	  }
-	  printf(" %4.1f", -10*log(erate/nseq)/log(10));
-	}
-      printf ("\n") ;
+        { nseq -= nlen[i] ;
+          printf ("\nbase %2d", i+1) ;
+          printf ("  %4.1f %4.1f %4.1f %4.1f %4.1f ",
+            100*((double)psum[i][_A]/nseq), 100*((double)psum[i][_C]/nseq),
+            100*((double)psum[i][_G]/nseq), 100*((double)psum[i][_T]/nseq),
+            100*((double)psum[i][_N]/nseq)) ;
+
+          for (erate = j = 0 ; j <= qMax ; ++j)
+            {
+              printf (" %3d",(int)lrint(1000*((double)pqsum[i][j]/nseq))) ;
+              erate += pow(10, j/-10.0) * pqsum[i][j];
+            }
+          printf(" %4.1f", -10*log(erate/nseq)/log(10));
+
+        }
+        printf ("\n") ;
     }
   exit(EXIT_SUCCESS);
-
 }
